# swagger.yaml
openapi: 3.0.0
info:
  title: UberEATS API
  description: API documentation for UberEATS prototype
  version: 1.0.0
servers:
  - url: http://localhost:3001/api
    description: Local development server
...
paths:
...
  /restaurants:
    get:
      tags:
        - Restaurant
      summary: Get all restaurants
      description: Get a list of all restaurants
      responses:
        '200':
          description: List of restaurants
        '500':
          description: Server error
  /restaurants/{id}:
    get:
      tags:
        - Restaurant
      summary: Get restaurant by ID
      description: Get details of a specific restaurant
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 2
      responses:
        '200':
          description: Restaurant details
        '404':
          description: Restaurant not found
        '500':
          description: Server error
  /restaurants/{id}/menu:
    get:
      tags:
        - Restaurant
      summary: Get restaurant menu
      description: Get the menu for a specific restaurant
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 2
      responses:
        '200':
          description: Restaurant menu
        '404':
          description: Restaurant not found
        '500':
          description: Server error
  /restaurant/profile:
    get:
      tags:
        - Restaurant
      summary: Get restaurant profile
      description: Get logged in restaurant's profile
      responses:
        '200':
          description: Restaurant profile data
        '401':
          description: Unauthorized
        '404':
          description: Profile not found
        '500':
          description: Server error
    put:
      tags:
        - Restaurant
      summary: Update restaurant profile
      description: Update logged in restaurant's profile information
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                description:
                  type: string
                  example: 'Authentic Italian cuisine'
                location:
                  type: string
                  example: '123 Main St, San Jose, CA'
                cuisine:
                  type: string
                  example: 'Italian'
                openTime:
                  type: string
                  format: time
                  example: '11:00'
                closeTime:
                  type: string
                  format: time
                  example: '22:00'
                deliveryFee:
                  type: number
                  example: 2.99
                minOrder:
                  type: number
                  example: 15
      responses:
        '200':
          description: Profile updated successfully
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '404':
          description: Profile not found
        '500':
          description: Server error
  /restaurant/profile/image:
    post:
      tags:
        - Restaurant
      summary: Upload restaurant image
      description: Upload an image for the restaurant
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                image:
                  type: string
                  format: binary
      responses:
        '200':
          description: Restaurant image updated successfully
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '404':
          description: Profile not found
        '500':
          description: Server error
  /restaurant/dishes:
    get:
      tags:
        - Restaurant
      summary: Get restaurant dishes
      description: Get all dishes for the logged in restaurant
      responses:
        '200':
          description: List of dishes
        '401':
          description: Unauthorized
        '500':
          description: Server error
    post:
      tags:
        - Restaurant
      summary: Add a new dish
      description: Add a new dish to the restaurant's menu
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - name
                - description
                - price
                - category
              properties:
                name:
                  type: string
                  example: 'Spaghetti Carbonara'
                description:
                  type: string
                  example: 'Classic Italian pasta with eggs, cheese, pancetta, and black pepper'
                price:
                  type: number
                  example: 14.99
                category:
                  type: string
                  example: 'Main Course'
                image:
                  type: string
                  format: binary
      responses:
        '201':
          description: Dish added successfully
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '500':
          description: Server error
  /restaurant/dishes/{id}:
    put:
      tags:
        - Restaurant
      summary: Update a dish
      description: Update an existing dish
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - name
                - description
                - price
                - category
              properties:
                name:
                  type: string
                  example: 'Spaghetti Carbonara'
                description:
                  type: string
                  example: 'Classic Italian pasta with eggs, cheese, pancetta, and black pepper'
                price:
                  type: number
                  example: 14.99
                category:
                  type: string
                  example: 'Main Course'
                isAvailable:
                  type: boolean
                  example: true
                image:
                  type: string
                  format: binary
      responses:
        '200':
          description: Dish updated successfully
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Dish not found
        '500':
          description: Server error
    delete:
      tags:
        - Restaurant
      summary: Delete a dish
      description: Delete an existing dish
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        '200':
          description: Dish deleted successfully
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Dish not found
        '500':
          description: Server error
  /restaurant/orders:
    get:
      tags:
        - Restaurant
        - Orders
      summary: Get restaurant orders
      description: Get orders for the logged in restaurant, optionally filtered by status
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [New, Order Received, Preparing, On the Way, Pick-up Ready, Delivered, Picked Up, Cancelled]
          example: Preparing
      responses:
        '200':
          description: List of orders
        '401':
          description: Unauthorized
        '500':
          description: Server error
  /restaurant/orders/{id}/status:
    put:
      tags:
        - Restaurant
        - Orders
      summary: Update order status
      description: Update the status of an order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - status
              properties:
                status:
                  type: string
                  enum: [New, Order Received, Preparing, On the Way, Pick-up Ready, Delivered, Picked Up, Cancelled]
                  example: Preparing
      responses:
        '200':
          description: Order status updated successfully
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Order not found
        '500':
          description: Server error
  /orders/{id}:
    get:
      tags:
        - Orders
      summary: Get order details
      description: Get details of a specific order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        '200':
          description: Order details
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Order not found
        '500':
          description: Server error
...